airdrop-discovery/research/requirements.txt
# Core requirements for web scraping and analysis
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.36.0
python-dotenv==1.0.0
robotexclusionrulesparser==1.7.1
markdown==3.4.4
pandas==2.0.3
aiohttp==3.8.5
asyncio==3.4.3
lxml==4.9.3
html5lib==1.1
tqdm==4.65.0
python-frontmatter==1.0.0
PyYAML==6.0.1
